+++ mealpy/utils/app/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python
+# Created by "Thieu" at 22:21, 06/09/2022 ----------%                                                                               
+#       Email: nguyenthieu2102@gmail.com            %                                                    
+#       Github: https://github.com/thieu1995        %                         
+# --------------------------------------------------%
+
+from .multitask import Multitask